Remove unused blake2 function to avoid undefined references
Compilers are not required to elide static functions which are unused.

Some compilers, such as Solaris Studio, always emits the function, even
if the function does not get called within the translation unit.  This
becomes problematic when a static inline function calls a non-existent
function; thus, resulting in (dynamic or static) link time errors.

Given that 'blake2' is never referenced nor called, remove the
definition of this function to increase portability for non-Linux
toolchains.

Signed-off-by: Eric N. Vander Weele <ericvw@gmail.com>