Simplify linking of shared libraries on the AIX OS
Have the approach of building shared libraries on the AIX operating
system be similar to that of a System V system.  The primary benefit of
this change is the elimination of custom AIX paths and reducing the
changes at `./configure` to affect just the `LDSHARED` environment
variable.

TODO: Highlight the primary difference in what `-G` controls.
TODO: Walk through existing options from various paths and explain why
they were they were either removed or changed.